refactor(core/ui): centralise unsigned-byte reads and drop vararg double casts
Cast audit cleanup, batch 2:

#2 - ImGui::Text vararg promotion is automatic, and
cppcoreguidelines-pro-type-vararg / hicpp-vararg are explicitly disabled in
.clang-tidy, so the static_cast<double>(float) casts on zoom_/pan_px_ were
pure noise. Removed 7 occurrences in RenderStatusBar/RenderMainUi.

#5 - Introduced core::ToUnsignedByte(char) in src/core/ByteOps.h to
deduplicate the static_cast<uint32_t>(static_cast<uint8_t>(c)) double-cast
that appears whenever we treat a raw char as an unsigned byte. Applied to:

- src/core/Base64.h: 5 sites in Base64Encode/Decode + the decode-table
  builder.
- src/ui/branch/DrawBranchTextOnPath.cpp: 7 sites in the UTF-8 codepoint
  decoder (NextUtf8Codepoint).

Net: ~20 cast occurrences gone across the two files; the remaining casts in
those files are intentional (int->size_t/ptrdiff_t for indexing, narrowing
back to char for byte output, int->ImWchar for ImGui glyph lookup) and stay.

Validated: scripts/build_tests_macos.sh (10/10 ctest passing, including
base64 round-trip and full-byte-range tests).

+++ src/ui/MindMapUi.cpp
@@ -309,13 +309,12 @@ void RenderStatusBar(const UiState& state) {
         "Status  Branches: %s  |  Edge to \"%s\": %s  |  Zoom: %.2f  |  Pan: (%.1f, %.1f)",
         state.canvas_.GetBranchStyleComboPreviewLabel(), edge_child_label,
         mind_map::ui::branch::GetBranchStyleDisplayName(state.canvas_.GetBranchStyleForSelectedChildEdge()),
-        static_cast<double>(state.zoom_), static_cast<double>(state.pan_px_.x),
-        static_cast<double>(state.pan_px_.y));
+        state.zoom_, state.pan_px_.x, state.pan_px_.y);
   }
   else {
     ImGui::Text("Status  Branches: %s  |  Selected edge: none  |  Zoom: %.2f  |  Pan: (%.1f, %.1f)",
-                state.canvas_.GetBranchStyleComboPreviewLabel(), static_cast<double>(state.zoom_),
-                static_cast<double>(state.pan_px_.x), static_cast<double>(state.pan_px_.y));
+                state.canvas_.GetBranchStyleComboPreviewLabel(), state.zoom_,
+                state.pan_px_.x, state.pan_px_.y);
   }
 }
 
@@ -434,7 +433,7 @@ void RenderMainUi(UiState& state, mind_map::app::DocumentSessionService& session
     ImGui::TextUnformatted(
         "Canvas: drag nodes. Drag empty space to pan. Mouse wheel zooms. Click a non-root node to edit its incoming "
         "edge style; root or empty canvas clears selection. \"Set all branches\" applies one style to every edge.");
-    ImGui::Text("Zoom %.2f", static_cast<double>(state.zoom_));
+    ImGui::Text("Zoom %.2f", state.zoom_);
     RenderCanvas(state, session);
   }
   ImGui::EndChild();
